fix eval

+++ utils/evaluate.py
@@ -50,7 +50,7 @@ def eval(valid_dataloader, model, norm_flag):
         # compute loss and acc for every video
         avg_single_video_output = sum(output_dict_tmp[key]) / len(output_dict_tmp[key])
         avg_single_video_target = sum(target_dict_tmp[key]) / len(target_dict_tmp[key])
-        loss = criterion(avg_single_video_output, avg_single_video_target)
+        loss = criterion(avg_single_video_output.long(), avg_single_video_target.long())
         acc_valid = accuracy(avg_single_video_output, avg_single_video_target, topk=(1,))
         valid_losses.update(loss.item())
         valid_top1.update(acc_valid[0])
